# docs/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Piums Platform API
  description: |
    API completa de la plataforma Piums - Marketplace para contratar artistas callejeros.
    
    ## Autenticación
    La mayoría de los endpoints requieren autenticación mediante JWT Bearer token.
    
    ```
    Authorization: Bearer <token>
    ```
    
    ## Ambientes
    - **Development**: http://localhost:3000
    - **Staging**: https://staging.piums.com
    - **Production**: https://piums.com
    
    ## Rate Limiting
    - **Login**: 5 requests / 15 minutos
    - **Register**: 3 requests / hora
    - **General**: 100 requests / 15 minutos
    
  version: 1.0.0
  contact:
    name: Piums API Support
    email: soporte@piums.io
    url: https://piums.com/docs
  license:
    name: Proprietary
    url: https://piums.com/terms
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://staging.piums.com
    description: Staging server
  - url: https://piums.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtenido del endpoint /api/auth/login
  schemas:
...
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    
    # ========================================================================
    # User Schemas
    # ========================================================================
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        role:
          type: string
          enum: [CLIENT, ARTIST, ADMIN]
        avatar:
          type: string
          format: uri
          nullable: true
          example: "https://res.cloudinary.com/piums/image/upload/v1234567890/avatars/user.jpg"
        phone:
          type: string
          nullable: true
          example: "+34600000000"
        emailVerified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        bio:
          type: string
        location:
          type: string
...
    Artist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        stageName:
          type: string
          example: "El Mago de la Calle"
        bio:
          type: string
        specialties:
          type: array
          items:
            type: string
          example: ["Música", "Malabares", "Magia"]
        location:
          type: object
          properties:
            city:
              type: string
              example: "Madrid"
            country:
              type: string
              example: "España"
            coordinates:
              type: object
              properties:
                lat:
                  type: number
                  example: 40.4168
                lng:
                  type: number
                  example: -3.7038
        rating:
          type: number
          format: float
          example: 4.8
        reviewCount:
          type: integer
          example: 127
        verified:
          type: boolean
          example: true
        portfolio:
          type: array
          items:
            type: string
            format: uri
        socialLinks:
          type: object
          properties:
            instagram:
              type: string
            youtube:
              type: string
            tiktok:
              type: string
        createdAt:
          type: string
          format: date-time
    
    # ========================================================================
    # Catalog Schemas
    # ========================================================================
...
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        artistId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        date:
          type: string
          format: date
          example: "2026-03-15"
        time:
          type: string
          example: "18:00"
        duration:
          type: integer
          example: 60
        location:
          type: object
          properties:
            address:
              type: string
            city:
              type: string
            postalCode:
              type: string
            coordinates:
              type: object
              properties:
                lat:
                  type: number
                lng:
                  type: number
        price:
          type: number
          format: float
          example: 150.00
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED, RESCHEDULED]
          example: "PENDING"
        paymentStatus:
          type: string
          enum: [PENDING, PAID, REFUNDED, FAILED]
        confirmationCode:
          type: string
          example: "BKG-ABC123"
        notes:
          type: string
        rescheduledAt:
          type: string
          format: date-time
          nullable: true
        rescheduleReason:
          type: string
          nullable: true
        rescheduleCount:
          type: integer
          example: 0
        createdAt:
          type: string
          format: date-time
    CreateBookingRequest:
      type: object
      required:
        - artistId
        - serviceId
        - date
        - time
        - location
      properties:
        artistId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        time:
          type: string
        location:
          type: object
          required:
            - address
            - city
          properties:
            address:
              type: string
            city:
              type: string
            postalCode:
              type: string
        notes:
          type: string
...
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        amount:
          type: number
          format: float
          example: 150.00
        currency:
          type: string
          example: "EUR"
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED, REFUNDED]
        paymentMethodId:
          type: string
          format: uuid
        stripePaymentIntentId:
          type: string
        createdAt:
          type: string
          format: date-time
    
    # ========================================================================
    # Review Schemas
    # ========================================================================
...
    CreateReviewRequest:
      type: object
      required:
        - bookingId
        - rating
      properties:
        bookingId:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        images:
          type: array
          items:
            type: string
            format: uri
    
    # ========================================================================
    # Notification Schemas
    # ========================================================================
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            total:
              type: integer
              example: 156
            pages:
              type: integer
              example: 8
paths:
...
  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Renovar token
      description: Obtener un nuevo access token usando el refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token renovado
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expiresIn:
                    type: string
        '401':
          description: Refresh token inválido
...
  /api/users/me:
    get:
      tags:
        - Users
      summary: Obtener perfil del usuario actual
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: No autenticado
    
    put:
      tags:
        - Users
      summary: Actualizar perfil
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Perfil actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    
    delete:
      tags:
        - Users
      summary: Eliminar cuenta
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Cuenta eliminada
...
  /api/artists:
    get:
      tags:
        - Artists
      summary: Listar artistas
      description: Obtener listado de artistas con paginación
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de artistas
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Artist'
    
    post:
      tags:
        - Artists
      summary: Crear perfil de artista
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - stageName
                - bio
                - specialties
              properties:
                stageName:
                  type: string
                bio:
                  type: string
                specialties:
                  type: array
                  items:
                    type: string
                location:
                  type: object
      responses:
        '201':
          description: Perfil de artista creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
...
  /api/bookings:
    get:
      tags:
        - Bookings
      summary: Listar reservas del usuario
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Lista de reservas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    
    post:
      tags:
        - Bookings
      summary: Crear reserva
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Reserva creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Datos inválidos o artista no disponible
...
  /api/payments:
    get:
      tags:
        - Payments
      summary: Historial de pagos
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Lista de pagos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    
    post:
      tags:
        - Payments
      summary: Procesar pago
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bookingId
                - paymentMethodId
              properties:
                bookingId:
                  type: string
                paymentMethodId:
                  type: string
      responses:
        '200':
          description: Pago procesado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'

  # ==========================================================================
  # REVIEWS
  # ==========================================================================
  /api/reviews:
    get:
      tags:
        - Reviews
      summary: Listar reseñas
      parameters:
        - name: artistId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Lista de reseñas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    
    post:
      tags:
        - Reviews
      summary: Crear reseña
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Reseña creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
...
  /api/notifications:
    get:
      tags:
        - Notifications
      summary: Listar notificaciones
      security:
        - bearerAuth: []
      parameters:
        - name: unread
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Lista de notificaciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
...
  /api/search/artists:
    get:
      tags:
        - Search
      summary: Buscar artistas
      description: Búsqueda avanzada de artistas con filtros
      parameters:
        - name: q
          in: query
          description: Término de búsqueda
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: priceMin
          in: query
          schema:
            type: number
        - name: priceMax
          in: query
          schema:
            type: number
        - name: rating
          in: query
          description: Rating mínimo
          schema:
            type: number
        - name: availability
          in: query
          description: Fecha de disponibilidad
          schema:
            type: string
            format: date
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance, price_asc, price_desc, rating, reviews]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Resultados de búsqueda
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
...